upstream.py: Change scorer for extractOne
Using fuzz.token_sort_ratio is much faster than the default scorer
for extractOne. Since we are using the result as guidance only and
do more detailed tests afterwards, faster is better. Lower the
minimum required score for further tests to ensure that nothing gets
lost. Also, use an even lower score cutoff. If a match is found
which is below the minimum acceptable score, display the best match
to guide with debugging.

The change reduces the time needed to handle the sample data (rebase
of chromeos-4.14 to v4.18-rc4) almost 4-fold from ~6m30s to ~1m45s.

While at it, improve regulatr expression for FIXUP: detection to accept
any number of whilespace after the FIXUP: tag, and permit _' as part of
words used for word hashing.

Signed-off-by: Guenter Roeck <groeck@chromium.org>