upstream.py: Limit patch fuzzy comparison size
Some incoming patches are extremely large; one patch with more than 100,000 (!)
lines was observed. This results in extremely long fuzzy match times.
Limit patch size to 2,000 lines. If one of the to-be-compared patches is larger,
take the patch size into account as well and abort if the difference is too
substantial.

Signed-off-by: Guenter Roeck <groeck@chromium.org>