update.py: Update 'usha' in rebase database
'usha' reflects a matching commit either in the upstream or in the -next
database. The initial recorded SHA is from a cherry-pick statement in the
commit message. Verify if it reflects an upstream commit. If it doesn't,
drop it. If there is a subject match but the SHA does not match, replace
the usha entry with the real SHA.

Signed-off-by: Guenter Roeck <groeck@chromium.org>