Determine rebase_baseline and stable_baseline automatically
We know that the rebase baseline is the merge point of
the Chrome OS branch, and we know that the stable baseline is the
most recent stable tag in the Chrome OS branch. This means that it is
not necessary to configure it, and having to configure it just introduces
potential problems. Determine it automatically.

Signed-off-by: Guenter Roeck <groeck@chromium.org>